Move ComponentSpec class and consolidate ComponentSpecs for official TFX components.
This change moves tfx.components.base.base_component.ComponentSpec and associated classes to tfx.types.component_spec.*; it also moves ComponentSpecs for official TFX components to tfx.types.standard_component_specs.*.

PiperOrigin-RevId: 262470181
