+++ tfx/components/base/base_component_test.py
@@ -17,18 +17,16 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import json
-
 import tensorflow as tf
 from tfx import types
 from tfx.components.base import base_component
 from tfx.components.base import base_executor
-from tfx.components.base.base_component import ChannelParameter
-from tfx.components.base.base_component import ExecutionParameter
 from tfx.proto import example_gen_pb2
+from tfx.types.component_spec import ChannelParameter
+from tfx.types.component_spec import ExecutionParameter
 
 
-class _BasicComponentSpec(base_component.ComponentSpec):
+class _BasicComponentSpec(types.ComponentSpec):
 
   PARAMETERS = {
       'folds': ExecutionParameter(type=int),
@@ -49,7 +47,7 @@ class _BasicComponent(base_component.BaseComponent):
   EXECUTOR_CLASS = base_executor.BaseExecutor
 
   def __init__(self,
-               spec: base_component.ComponentSpec = None,
+               spec: types.ComponentSpec = None,
                folds: int = None,
                input: types.Channel = None):  # pylint: disable=redefined-builtin
     if not spec:
@@ -57,195 +55,6 @@ class _BasicComponent(base_component.BaseComponent):
       spec = _BasicComponentSpec(
           folds=folds, input=input, output=output)
     super(_BasicComponent, self).__init__(spec=spec)
-
-
-class ComponentSpecTest(tf.test.TestCase):
-
-  def testComponentspecEmpty(self):
-
-    class EmptyComponentSpec(base_component.ComponentSpec):
-      PARAMETERS = {}
-      INPUTS = {}
-      OUTPUTS = {}
-
-    _ = EmptyComponentSpec()
-
-  def testComponentspecBasic(self):
-    proto = example_gen_pb2.Input()
-    proto.splits.extend([
-        example_gen_pb2.Input.Split(name='name1', pattern='pattern1'),
-        example_gen_pb2.Input.Split(name='name2', pattern='pattern2'),
-        example_gen_pb2.Input.Split(name='name3', pattern='pattern3'),])
-    input_channel = types.Channel(type_name='InputType')
-    output_channel = types.Channel(type_name='OutputType')
-    spec = _BasicComponentSpec(folds=10,
-                               proto=proto,
-                               input=input_channel,
-                               output=output_channel)
-    # Verify proto property.
-    self.assertIsInstance(spec.exec_properties['proto'], str)
-    decoded_proto = json.loads(spec.exec_properties['proto'])
-    self.assertCountEqual(['splits'], decoded_proto.keys())
-    self.assertEqual(3, len(decoded_proto['splits']))
-    self.assertCountEqual(
-        ['name1', 'name2', 'name3'],
-        list(s['name'] for s in decoded_proto['splits']))
-    self.assertCountEqual(
-        ['pattern1', 'pattern2', 'pattern3'],
-        list(s['pattern'] for s in decoded_proto['splits']))
-
-    # Verify other properties.
-    self.assertEqual(10, spec.exec_properties['folds'])
-    self.assertIs(spec.inputs.input, input_channel)
-    self.assertIs(spec.outputs.output, output_channel)
-
-    with self.assertRaisesRegexp(
-        TypeError,
-        "Expected type <(class|type) 'int'> for parameter 'folds' but got "
-        "string."):
-      spec = _BasicComponentSpec(folds='string', input=input_channel,
-                                 output=output_channel)
-
-    with self.assertRaisesRegexp(
-        TypeError,
-        'Expected InputType but found WrongType'):
-      spec = _BasicComponentSpec(
-          folds=10,
-          input=types.Channel(type_name='WrongType'),
-          output=output_channel)
-
-    with self.assertRaisesRegexp(
-        TypeError,
-        'Expected OutputType but found WrongType'):
-      spec = _BasicComponentSpec(
-          folds=10,
-          input=input_channel,
-          output=types.Channel(type_name='WrongType'))
-
-  def testInvalidComponentspecMissingProperties(self):
-
-    with self.assertRaisesRegexp(TypeError, "Can't instantiate abstract class"):
-      class InvalidComponentSpecA(base_component.ComponentSpec):
-        # Missing PARAMETERS.
-        INPUTS = {}
-        OUTPUTS = {}
-
-      InvalidComponentSpecA()
-
-    with self.assertRaisesRegexp(TypeError, "Can't instantiate abstract class"):
-      class InvalidComponentSpecB(base_component.ComponentSpec):
-        PARAMETERS = {}
-        # Missing INPUTS.
-        OUTPUTS = {}
-
-      InvalidComponentSpecB()
-
-    with self.assertRaisesRegexp(TypeError, "Can't instantiate abstract class"):
-      class InvalidComponentSpecC(base_component.ComponentSpec):
-        PARAMETERS = {}
-        INPUTS = {}
-        # Missing OUTPUTS.
-
-      InvalidComponentSpecC()
-
-  def testInvalidComponentspecWrongProperties(self):
-
-    with self.assertRaisesRegexp(TypeError,
-                                 'must override PARAMETERS with a dict'):
-      class InvalidComponentSpecA(base_component.ComponentSpec):
-        PARAMETERS = object()
-        INPUTS = {}
-        OUTPUTS = {}
-
-      InvalidComponentSpecA()
-
-    with self.assertRaisesRegexp(TypeError,
-                                 'must override INPUTS with a dict'):
-      class InvalidComponentSpecB(base_component.ComponentSpec):
-        PARAMETERS = {}
-        INPUTS = object()
-        OUTPUTS = {}
-
-      InvalidComponentSpecB()
-
-    with self.assertRaisesRegexp(TypeError,
-                                 'must override OUTPUTS with a dict'):
-      class InvalidComponentSpecC(base_component.ComponentSpec):
-        PARAMETERS = {}
-        INPUTS = {}
-        OUTPUTS = object()
-
-      InvalidComponentSpecC()
-
-  def testInvalidComponentspecWrongType(self):
-
-    class WrongTypeComponentSpecA(base_component.ComponentSpec):
-      PARAMETERS = {'x': object()}
-      INPUTS = {}
-      OUTPUTS = {}
-
-    with self.assertRaisesRegexp(ValueError,
-                                 'expects .* dicts are _ComponentParameter'):
-      _ = WrongTypeComponentSpecA()
-
-    class WrongTypeComponentSpecB(base_component.ComponentSpec):
-      PARAMETERS = {'x': ChannelParameter(type_name='X')}
-      INPUTS = {}
-      OUTPUTS = {}
-
-    with self.assertRaisesRegexp(TypeError,
-                                 'expects values of type ExecutionParameter'):
-      _ = WrongTypeComponentSpecB()
-
-    class WrongTypeComponentSpecC(base_component.ComponentSpec):
-      PARAMETERS = {}
-      INPUTS = {'x': ExecutionParameter(type=int)}
-      OUTPUTS = {}
-
-    with self.assertRaisesRegexp(TypeError,
-                                 'expect values of type ChannelParameter'):
-      _ = WrongTypeComponentSpecC()
-
-    class WrongTypeComponentSpecD(base_component.ComponentSpec):
-      PARAMETERS = {}
-      INPUTS = {'x': ExecutionParameter(type=int)}
-      OUTPUTS = {}
-
-    with self.assertRaisesRegexp(TypeError,
-                                 'expect values of type ChannelParameter'):
-      _ = WrongTypeComponentSpecD()
-
-  def testInvalidComponentspecDuplicateProperty(self):
-
-    class DuplicatePropertyComponentSpec(base_component.ComponentSpec):
-      PARAMETERS = {'x': ExecutionParameter(type=int)}
-      INPUTS = {'x': ChannelParameter(type_name='X')}
-      OUTPUTS = {}
-
-    with self.assertRaisesRegexp(ValueError,
-                                 'has a duplicate argument'):
-      _ = DuplicatePropertyComponentSpec()
-
-  def testComponentspecMissingArguments(self):
-
-    class SimpleComponentSpec(base_component.ComponentSpec):
-      PARAMETERS = {
-          'x': ExecutionParameter(type=int),
-          'y': ExecutionParameter(type=int, optional=True),
-      }
-      INPUTS = {'z': ChannelParameter(type_name='Z')}
-      OUTPUTS = {}
-
-    with self.assertRaisesRegexp(ValueError,
-                                 'Missing argument'):
-      _ = SimpleComponentSpec(x=10)
-
-    with self.assertRaisesRegexp(ValueError,
-                                 'Missing argument'):
-      _ = SimpleComponentSpec(z=types.Channel(type_name='Z'))
-
-    # Okay since y is optional.
-    _ = SimpleComponentSpec(x=10, z=types.Channel(type_name='Z'))
 
 
 class ComponentTest(tf.test.TestCase):
@@ -262,7 +71,7 @@ class ComponentTest(tf.test.TestCase):
 
     with self.assertRaisesRegexp(
         ValueError,
-        'expects "spec" argument to be an instance of ComponentSpec'):
+        'expects "spec" argument to be an instance of types.ComponentSpec'):
       _ = _BasicComponent(spec=object())  # pytype: disable=wrong-arg-types
 
   def testComponentSpecClass(self):
@@ -277,9 +86,8 @@ class ComponentTest(tf.test.TestCase):
       MissingSpecComponent(spec=object())  # pytype: disable=wrong-arg-types
 
     with self.assertRaisesRegexp(
-        TypeError,
-        'expects SPEC_CLASS property to be a subclass of '
-        'base_component.ComponentSpec'):
+        TypeError, 'expects SPEC_CLASS property to be a subclass of '
+        'types.ComponentSpec'):
       MissingSpecComponent._validate_component_class()
 
     class InvalidSpecComponent(base_component.BaseComponent):
@@ -288,9 +96,8 @@ class ComponentTest(tf.test.TestCase):
       EXECUTOR_CLASS = base_executor.BaseExecutor
 
     with self.assertRaisesRegexp(
-        TypeError,
-        'expects SPEC_CLASS property to be a subclass of '
-        'base_component.ComponentSpec'):
+        TypeError, 'expects SPEC_CLASS property to be a subclass of '
+        'types.ComponentSpec'):
       InvalidSpecComponent._validate_component_class()
 
   def testComponentExecutorClass(self):
@@ -323,7 +130,7 @@ class ComponentTest(tf.test.TestCase):
 
   def testComponentCustomExecutor(self):
 
-    class EmptyComponentSpec(base_component.ComponentSpec):
+    class EmptyComponentSpec(types.ComponentSpec):
       PARAMETERS = {}
       INPUTS = {}
       OUTPUTS = {}
