+++ tfx/orchestration/airflow/airflow_dag_runner_test.py
@@ -25,54 +25,55 @@ from tfx.components.base import base_component
 from tfx.components.base import base_executor
 from tfx.orchestration import pipeline
 from tfx.orchestration.airflow import airflow_dag_runner
+from tfx.types import component_spec
 
 
-class _FakeComponentSpecA(base_component.ComponentSpec):
+class _FakeComponentSpecA(types.ComponentSpec):
   PARAMETERS = {}
   INPUTS = {}
-  OUTPUTS = {'output': base_component.ChannelParameter(type_name='a')}
+  OUTPUTS = {'output': component_spec.ChannelParameter(type_name='a')}
 
 
-class _FakeComponentSpecB(base_component.ComponentSpec):
+class _FakeComponentSpecB(types.ComponentSpec):
   PARAMETERS = {}
-  INPUTS = {'a': base_component.ChannelParameter(type_name='a')}
-  OUTPUTS = {'output': base_component.ChannelParameter(type_name='b')}
+  INPUTS = {'a': component_spec.ChannelParameter(type_name='a')}
+  OUTPUTS = {'output': component_spec.ChannelParameter(type_name='b')}
 
 
-class _FakeComponentSpecC(base_component.ComponentSpec):
-  PARAMETERS = {}
-  INPUTS = {
-      'a': base_component.ChannelParameter(type_name='a'),
-      'b': base_component.ChannelParameter(type_name='b')
-  }
-  OUTPUTS = {'output': base_component.ChannelParameter(type_name='c')}
-
-
-class _FakeComponentSpecD(base_component.ComponentSpec):
+class _FakeComponentSpecC(types.ComponentSpec):
   PARAMETERS = {}
   INPUTS = {
-      'b': base_component.ChannelParameter(type_name='b'),
-      'c': base_component.ChannelParameter(type_name='c'),
+      'a': component_spec.ChannelParameter(type_name='a'),
+      'b': component_spec.ChannelParameter(type_name='b')
   }
-  OUTPUTS = {'output': base_component.ChannelParameter(type_name='d')}
+  OUTPUTS = {'output': component_spec.ChannelParameter(type_name='c')}
 
 
-class _FakeComponentSpecE(base_component.ComponentSpec):
+class _FakeComponentSpecD(types.ComponentSpec):
   PARAMETERS = {}
   INPUTS = {
-      'a': base_component.ChannelParameter(type_name='a'),
-      'b': base_component.ChannelParameter(type_name='b'),
-      'd': base_component.ChannelParameter(type_name='d'),
+      'b': component_spec.ChannelParameter(type_name='b'),
+      'c': component_spec.ChannelParameter(type_name='c'),
   }
-  OUTPUTS = {'output': base_component.ChannelParameter(type_name='e')}
+  OUTPUTS = {'output': component_spec.ChannelParameter(type_name='d')}
+
+
+class _FakeComponentSpecE(types.ComponentSpec):
+  PARAMETERS = {}
+  INPUTS = {
+      'a': component_spec.ChannelParameter(type_name='a'),
+      'b': component_spec.ChannelParameter(type_name='b'),
+      'd': component_spec.ChannelParameter(type_name='d'),
+  }
+  OUTPUTS = {'output': component_spec.ChannelParameter(type_name='e')}
 
 
 class _FakeComponent(base_component.BaseComponent):
 
-  SPEC_CLASS = base_component.ComponentSpec
+  SPEC_CLASS = types.ComponentSpec
   EXECUTOR_CLASS = base_executor.BaseExecutor
 
-  def __init__(self, spec: base_component.ComponentSpec):
+  def __init__(self, spec: types.ComponentSpec):
     component_name = spec.__class__.__name__.replace(
         '_FakeComponentSpec', '_FakeComponent')
     super(_FakeComponent, self).__init__(
