+++ tfx/orchestration/interactive/interactive_context_test.py
@@ -24,13 +24,14 @@ from tfx import types
 from tfx.components.base import base_component
 from tfx.components.base import base_executor
 from tfx.orchestration.interactive import interactive_context
+from tfx.types import component_spec
 
 
 class InteractiveContextTest(tf.test.TestCase):
 
   def testBasicRun(self):
 
-    class _FakeComponentSpec(base_component.ComponentSpec):
+    class _FakeComponentSpec(types.ComponentSpec):
       COMPONENT_NAME = '_FakeComponent'
       PARAMETERS = {}
       INPUTS = {}
@@ -48,7 +49,7 @@ class InteractiveContextTest(tf.test.TestCase):
       SPEC_CLASS = _FakeComponentSpec
       EXECUTOR_CLASS = _FakeExecutor
 
-      def __init__(self, spec: base_component.ComponentSpec):
+      def __init__(self, spec: types.ComponentSpec):
         super(_FakeComponent, self).__init__(spec=spec)
 
     c = interactive_context.InteractiveContext()
@@ -58,12 +59,10 @@ class InteractiveContextTest(tf.test.TestCase):
 
   def testUnresolvedChannel(self):
 
-    class _FakeComponentSpec(base_component.ComponentSpec):
+    class _FakeComponentSpec(types.ComponentSpec):
       COMPONENT_NAME = '_FakeComponent'
       PARAMETERS = {}
-      INPUTS = {
-          'input': base_component.ChannelParameter(type_name='Foo')
-      }
+      INPUTS = {'input': component_spec.ChannelParameter(type_name='Foo')}
       OUTPUTS = {}
 
     class _FakeExecutor(base_executor.BaseExecutor):
@@ -78,7 +77,7 @@ class InteractiveContextTest(tf.test.TestCase):
       SPEC_CLASS = _FakeComponentSpec
       EXECUTOR_CLASS = _FakeExecutor
 
-      def __init__(self, spec: base_component.ComponentSpec):
+      def __init__(self, spec: types.ComponentSpec):
         super(_FakeComponent, self).__init__(spec=spec)
 
     c = interactive_context.InteractiveContext()
