+++ tfx/types/__init__.py
@@ -16,3 +16,4 @@
 from tfx.types.artifact import Artifact
 from tfx.types.artifact import DEFAULT_EXAMPLE_SPLITS
 from tfx.types.channel import Channel
+from tfx.types.component_spec import ComponentSpec
